Add static_cast to avoid conversion warning
I guess we're in trouble if we add any block ciphers with size >= 2^31
Probably fine.

+++ src/lib/ffi/ffi_block.cpp
@@ -59,7 +59,8 @@ int botan_block_cipher_set_key(botan_block_cipher_t bc,
 */
 int botan_block_cipher_block_size(botan_block_cipher_t bc)
    {
-   return BOTAN_FFI_DO(Botan::BlockCipher, bc, b, { return b.block_size(); });
+   return BOTAN_FFI_DO(Botan::BlockCipher, bc, b,
+                       { return static_cast<int>(b.block_size()); });
    }
 
 int botan_block_cipher_encrypt_blocks(botan_block_cipher_t bc,
